Desarrollo Api usuarios

+++ CRUD-USUARIO/usuarios/src/main/java/ecoMarket/duoc/cl/usuarios/service/UsuarioService.java
@@ -6,6 +6,7 @@ import ecoMarket.duoc.cl.usuarios.repository.UsuarioRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import javax.persistence.EntityNotFoundException;
 import javax.transaction.Transactional;
 import java.util.*;
 
@@ -15,20 +16,34 @@ import java.util.*;
 public class UsuarioService {
     @Autowired
     private UsuarioRepository usuarioRepository;
+
     public List<Usuario> findAll() {
         return usuarioRepository.findAll();
     }
     public Usuario findById(Integer id) {
-        return usuarioRepository.findById(id)
-                .orElseThrow(() -> new NoSuchElementException("Usuario no encontrado"));
-
+        if (!usuarioRepository.existsById(id)) {
+            throw new EntityNotFoundException("Usuario no encontrado");
+        }
+        return  usuarioRepository.findById(id).get();
     }
     public Usuario save(Usuario usuario) {
+        if (usuarioRepository.existsByRut(usuario.getRut())) {
+            throw new IllegalArgumentException("RUT ya registrado");
+        }
+
+        if (usuarioRepository.existsByEmail(usuario.getEmail())) {
+            throw new IllegalArgumentException("Email ya registrado");
+        }
+        if (usuarioRepository.existsByTelefono(usuario.getTelefono())) {
+            throw new IllegalArgumentException("Teléfono ya registrado");
+        }
         return usuarioRepository.save(usuario);
     }
-    public String deleteById(Integer id) {
+    public void deleteById(Integer id) {
+        if (!usuarioRepository.existsById(id)) {
+            throw new EntityNotFoundException("Usuario no encontrado");
+        }
         usuarioRepository.deleteById(id);
-        return "Usuario eliminado con éxito";
     }
     public Usuario update(Integer id, Usuario usuario) {
         Usuario usuario1 = findById(id);
@@ -42,6 +57,5 @@ public class UsuarioService {
         usuario1.setTelefono(usuario.getTelefono());
         return usuarioRepository.save(usuario1);
     }
-
 
 }
